ftp+socket

+++ ravpy/distributed/compute.py
@@ -16,8 +16,6 @@ from ..config import FTP_DOWNLOAD_FILES_FOLDER
 from ravop import functions
 import ast
 
-outputs = g.outputs
-ops = g.ops
 
 numpy_functions = {
             "neg": "np.negative",
@@ -117,7 +115,6 @@ def compute_locally_bm(*args, **kwargs):
 
 # async 
 def compute_locally(payload):
-    global outputs
 
     # print("Computing ",payload["operator"])
     # print('\n\nPAYLOAD: ',payload)
@@ -128,28 +125,34 @@ def compute_locally(payload):
     for i in range(len(payload["values"])):
         if "value" in payload["values"][i].keys():
             # print("From server")
-            server_file_path = payload["values"][i]["path"]
+            if "path" not in payload["values"][i].keys():
+                values.append(payload["values"][i]["value"])
+            
+            else:
+                server_file_path = payload["values"][i]["path"]
 
-            download_path = os.path.join(FTP_DOWNLOAD_FILES_FOLDER,os.path.basename(payload["values"][i]["path"]))
+                download_path = os.path.join(FTP_DOWNLOAD_FILES_FOLDER,os.path.basename(payload["values"][i]["path"]))
 
+                try:
+                    g.ftp_client.download(download_path, os.path.basename(server_file_path))
+                except Exception as error:
+                    print('Error: ', error)
+                    emit_error(payload, error)
 
-            g.ftp_client.download(download_path, os.path.basename(server_file_path))
+                value = load_data(download_path).tolist()
+                print('Loaded Data Value: ',value)
+                values.append(value)
 
+                if os.path.basename(server_file_path) not in g.delete_files_list and payload["values"][i]["to_delete"] == 'True':
+                    g.delete_files_list.append(os.path.basename(server_file_path))
 
-            value = load_data(download_path).tolist()
-            print('Loaded Data Value: ',value)
-            values.append(value)
-
-            if os.path.basename(server_file_path) not in g.delete_files_list and payload["values"][i]["to_delete"] == 'True':
-                g.delete_files_list.append(os.path.basename(server_file_path))
-
-            if os.path.exists(download_path):
-                os.remove(download_path)
+                if os.path.exists(download_path):
+                    os.remove(download_path)
 
         elif "op_id" in payload["values"][i].keys():
             # print("From client")
             try:
-                values.append(outputs[payload['values'][i]['op_id']])
+                values.append(g.outputs[payload['values'][i]['op_id']])
             except Exception as e:
                 emit_error(payload,e)
 
@@ -183,24 +186,45 @@ def compute_locally(payload):
 
 
             result = eval(expression)
-        # print("Result : \n",result)
-        file_path = upload_result(payload, result)
+        
+        result_byte_size = result.size * result.itemsize
 
-        outputs[payload["op_id"]] = result.tolist()
+        if result_byte_size < (10 * 1000000)//50:
+            try:
+                result = result.tolist()
+            except:
+                result = result
+        
+            g.outputs[payload["op_id"]] = result
 
-        op = ops[payload["op_id"]]
-        op["status"] = "success"
-        op["endTime"] = int(time.time() * 1000)
-        ops[payload["op_id"]] = op
-
-        return json.dumps({
+            return json.dumps({
             'op_type': payload["op_type"],
-            'file_name': os.path.basename(file_path),
+            'result': result,
             'username': g.cid,
             'operator': payload["operator"],
             "op_id": payload["op_id"],
             "status": "success"
-        })
+            })
+
+        else:        
+
+            file_path = upload_result(payload, result)
+
+            g.outputs[payload["op_id"]] = result.tolist()
+
+            # op = g.ops[payload["op_id"]]
+            # op["status"] = "success"
+            # op["endTime"] = int(time.time() * 1000)
+            # g.ops[payload["op_id"]] = op
+
+            return json.dumps({
+                'op_type': payload["op_type"],
+                'file_name': os.path.basename(file_path),
+                'username': g.cid,
+                'operator': payload["operator"],
+                "op_id": payload["op_id"],
+                "status": "success"
+            })
 
     except Exception as error:
         print('Error: ', error)
@@ -208,7 +232,6 @@ def compute_locally(payload):
 
 
 def upload_result(payload, result):
-    global outputs, ops
     try:
         result = result.tolist()
     except:
@@ -217,8 +240,8 @@ def upload_result(payload, result):
     # print("Emit Success")
 
     file_path = dump_data(payload['op_id'],result)
-
     g.ftp_client.upload(file_path, os.path.basename(file_path))
+    
     print("\nFile uploaded!", file_path)
     os.remove(file_path)
   
@@ -230,7 +253,6 @@ def emit_error(payload, error):
     # print(payload)
     print(error)
     error=str(error)
-    global ops
     client = g.client
     print(error,payload)
     client.emit("op_completed", json.dumps({
@@ -241,12 +263,30 @@ def emit_error(payload, error):
             "status": "failure"
     }), namespace="/client")
 
-    op = ops[payload["op_id"]]
-    op["status"] = "failure"
-    op["endTime"] = int(time.time() * 1000)
-    ops[payload["op_id"]] = op
+    # op = g.ops[payload["op_id"]]
+    # op["status"] = "failure"
+    # op["endTime"] = int(time.time() * 1000)
+    # g.ops[payload["op_id"]] = op
 
+    try:
+        for ftp_file in g.delete_files_list:
+            g.ftp_client.delete_file(ftp_file)
+    except Exception as e:
+        client.emit("op_completed", json.dumps({
+            'op_type': payload["op_type"],
+            'error': error,
+            'operator': payload["operator"],
+            "op_id": payload["op_id"],
+            "status": "failure"
+            }), namespace="/client")
 
+        g.delete_files_list = []
+        g.outputs = {}
+        g.has_subgraph = False
+
+    g.delete_files_list = []
+    g.outputs = {}
+    g.has_subgraph = False
 
 
 #ops:
